Fix docker build workflow failed when requirements.txt is empty (#6342)

+++ ironbank/build_ironbank_docker.sh
@@ -75,7 +75,16 @@ function build_hardening_manifest {
     docker pull $DOCKER_IMAGE
     DOCKER_PACKAGES_METADATA_PATH="$OUTPUT_PATH/docker_packages_metadata.txt"
     REQUIREMENTS="$(cat $1/requirements.txt)"
-    docker run -it $DOCKER_IMAGE /bin/sh -c "cd ~;dnf install -y --nodocs python$PYTHON_VERSION-devel gcc gcc-c++ make wget git;touch /requirements.txt;echo \"$REQUIREMENTS\" > /requirements.txt;pip uninstall -y -r /requirements.txt;pip cache purge;pip install -v --no-deps --no-cache-dir --log /tmp/pip.log -r /requirements.txt;cat /tmp/pip.log;exit" | grep Added >> $DOCKER_PACKAGES_METADATA_PATH
+
+    # Run the base image docker container only when requirements.txt exists
+    if [[ -f $REQUIREMENTS ]] && [[ -s $REQUIREMENTS ]]; then
+      echo "Prepare to Run the image docker container"
+      docker run -it $DOCKER_IMAGE /bin/sh -c "cd ~;dnf install -y --nodocs python$PYTHON_VERSION-devel gcc gcc-c++ make wget git;touch /requirements.txt;echo \"$REQUIREMENTS\" > /requirements.txt;pip uninstall -y -r /requirements.txt;pip cache purge;pip install -v --no-deps --no-cache-dir --log /tmp/pip.log -r /requirements.txt;cat /tmp/pip.log;exit" | grep Added >> $DOCKER_PACKAGES_METADATA_PATH
+    else
+      echo "Skip docker run - requirements.txt file is missing"
+    fi
+
+    echo "Prepare to build hardening_manifest.yaml"
     python ./ironbank/build_hardening_manifest.py --docker_image_dir $1 --output_path $OUTPUT_PATH --docker_packages_metadata_path $DOCKER_PACKAGES_METADATA_PATH
   else
     echo "Could not login to $REGISTRYONE_URL, aborting..."
@@ -86,6 +95,14 @@ function build_hardening_manifest {
 # $1: docker image dir (~/../docker/$IMAGE_NAME)
 function build_dockerfile {
   OUTPUT_PATH=ironbank/$(basename $1)
+  REQUIREMENTS="$OUTPUT_PATH/docker_packages_metadata.txt"
+
+  if [[ ! -f $REQUIREMENTS ]] && [[ ! -f $1/Dockerfile.ironbank ]]; then
+    echo "docker_packages_metadata.txt is missing in this docker, please create Dockerfile.ironbank, aborting..."
+    return 1;
+  fi
+
+
   if [[ ! -d $OUTPUT_PATH ]]; then
     mkdir $OUTPUT_PATH
   fi
@@ -116,7 +133,10 @@ function upload_image_to_artifacts {
   SOURCE_PATH="ironbank/$IMAGE_NAME"
   cp -r $SOURCE_PATH $TARGET_PATH
   cp $CURRENT_DIR/docker/$IMAGE_NAME/requirements.txt $TARGET_PATH
-  rm $SOURCE_PATH/docker_packages_metadata.txt
+  if [[ -f $SOURCE_PATH/docker_packages_metadata.txt ]]; then
+    rm $SOURCE_PATH/docker_packages_metadata.txt
+  fi
+
 }
 
 # $1: docker image dir (~/../docker/$IMAGE_NAME)
